# internal/pkg/docs/swagger.yaml
definitions:
  delivery.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  models.ProfileUpdatePayload:
    properties:
      description:
        type: string
      password:
        $ref: '#/definitions/models.passwords'
    type: object
  models.SignUpPayloadForSwagger:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
...
  models.UserForSwagger:
    properties:
      create_time:
        type: string
      description:
        type: string
      id:
        type: string
      image_path:
        type: string
      second_factor:
        type: boolean
      username:
        type: string
    type: object
  models.UserFormData:
    properties:
      code:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  models.passwords:
    properties:
      new:
        type: string
      old:
        type: string
    type: object
host: you-note.ru
info:
  contact: {}
  description: API for YouNote service
  title: YouNote API
  version: "1.0"
paths:
  /api/auth/check_user:
    get:
      description: Get user info if user is authorized
      operationId: check-user
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
      summary: Check user
      tags:
      - auth
  /api/auth/disable_2fa:
    delete:
      description: Remove secret for QR-code from database
      operationId: disable-second-factor
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
      summary: Disable second factor
      tags:
      - auth
  /api/auth/get_qr:
    get:
      description: Generate QR code for 2FA
      operationId: get-qr-code
      produces:
      - image/png
      responses:
        "200":
          description: QR-code
          schema:
            type: file
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
      summary: Get QR code
      tags:
      - auth
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Login as a user
      operationId: sign-in
      parameters:
      - description: login data
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.UserFormData'
      produces:
      - application/json
      responses:
        "200":
          description: user
          schema:
            $ref: '#/definitions/models.UserForSwagger'
        "202":
          description: Accepted
        "400":
          description: error
          schema:
            $ref: '#/definitions/delivery.ErrorResponse'
        "401":
          description: error
          schema:
            $ref: '#/definitions/delivery.ErrorResponse'
      summary: Sign in
      tags:
      - auth
  /api/auth/logout:
    delete:
      description: Quit from user`s account
      operationId: log-out
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
      summary: Log out
      tags:
      - auth
  /api/auth/signup:
    post:
      consumes:
      - application/json
      description: Add a new user to the database
      operationId: sign-up
      parameters:
      - description: registration data
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.SignUpPayloadForSwagger'
      produces:
      - application/json
      responses:
        "200":
          description: user
          schema:
            $ref: '#/definitions/models.UserForSwagger'
        "400":
          description: error
          schema:
            $ref: '#/definitions/delivery.ErrorResponse'
      summary: Sign up
      tags:
      - auth
...
  /api/profile/get:
    get:
      description: Get user info if user is authorized
      operationId: get-profile
      produces:
      - application/json
      responses:
        "200":
          description: user
          schema:
            $ref: '#/definitions/models.UserForSwagger'
        "401":
          description: Unauthorized
      summary: Get profile
      tags:
      - profile
  /api/profile/update:
    post:
      consumes:
      - application/json
      description: Change password and/or description
      operationId: update-profile
      parameters:
      - description: update data
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.ProfileUpdatePayload'
      produces:
      - application/json
      responses:
        "200":
          description: user
          schema:
            $ref: '#/definitions/models.UserForSwagger'
        "400":
          description: error
          schema:
            $ref: '#/definitions/delivery.ErrorResponse'
        "401":
          description: Unauthorized
      summary: Update profile
      tags:
      - profile
  /api/profile/update_avatar:
    post:
      consumes:
      - multipart/form-data
      description: Change sources
      operationId: update-profile-sources
      parameters:
      - description: avatar
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: user
          schema:
            $ref: '#/definitions/models.UserForSwagger'
        "400":
          description: error
          schema:
            $ref: '#/definitions/delivery.ErrorResponse'
        "401":
          description: Unauthorized
        "413":
          description: Request Entity Too Large
      summary: Update profile sources
      tags:
      - profile
swagger: "2.0"